Create SystemLocaleProvider
